Completed styling and ready for pull request
