--- finale/src/components/navbar2.jsx
@@ -1,18 +1,0 @@
-import React from 'react';
-
-const Navbar = () => {
-  return (
-    <div className="header">
-      <span className="logo_left">Logo</span>
-      <span className="title_left">Reports</span>
-      <div className="header_right">
-        <span className="active">Overall</span>
-        <span className>Specific</span>
-      </div>
-    </div>
-  );
-}
-
-export default Navbar;
-
-
