+++ finale/src/components/state.jsx
@@ -1,20 +1,22 @@
 import React, { Component } from 'react';
-import District from './district';
-import data from "../data"
-import dl_logo from "../img/download_icon.png"
 import TableHeader from './tableHeader';
+import District from './district';
+import data from "../data";
+import dl_logo from "../img/download_icon.png";
 
 class State extends Component {
   constructor() {
     super();
     this.state = {
-      expandedRows: []
+      expandedRows: [],
+      isHidden: true,
     };
   }
 
   //Click handler for expanding rows
   handleRowClick(rowId) {
     const currentExpandedRows = this.state.expandedRows;
+    console.log(currentExpandedRows);
 
     //If false that means that there is no district expanded row at the time of this click handler
     const isRowCurrentlyExpanded = currentExpandedRows.includes(rowId);
@@ -24,13 +26,12 @@ class State extends Component {
       currentExpandedRows.filter(id => id !== rowId) :
       currentExpandedRows.concat(rowId);
 
-    console.log(newExpandedRows);
-
+    this.setState({ isHidden: !this.state.isHidden });
     this.setState({ expandedRows: newExpandedRows });
   }
 
   //Rendering function for district row
-  renderItem(sta) {
+  renderRows(sta) {
 
     //id is the callback and handleRowClick is function that accepts callback to expand or collapse a row
     const clickCallback = () => this.handleRowClick(sta.id);
@@ -38,7 +39,7 @@ class State extends Component {
     const stateRow = [
       <tr className="stateRow" key={"state-row-data-" + sta.id}>
         <td>
-          <span className="capital Sta">S</span>
+          <span className="capital">S</span>
           <span>{sta.title}</span>
           <img className="dl_logo" src={dl_logo} alt="dl_icon" />
 
@@ -46,7 +47,7 @@ class State extends Component {
             <button
               className="toggle-btn"
               onClick={clickCallback}>{sta.subRegions.length} Districts
-              {this.state.expandedRows > 0 ? " -" : " +"}
+               {this.state.isHidden ? " +" : " -"}
             </button>
           }
         </td>
@@ -72,7 +73,7 @@ class State extends Component {
     let allItemRows = [];
 
     data.forEach(sta => {
-      const perStateRow = this.renderItem(sta);
+      const perStateRow = this.renderRows(sta);
       allItemRows = allItemRows.concat(perStateRow);
     });
